requirements.txt
Django==2.1.4
django-extensions==2.1.4
django-macaddress==1.5.0
netaddr==0.7.19
pkg-resources==0.0.0
pydotplus==2.0.2
pyparsing==2.3.0
pytz==2018.7
six==1.12.0
Werkzeug==0.14.1
